changed encounter chance to a float value

+++ Room.h
@@ -8,7 +8,7 @@ class Room {
 private:
     unsigned int id;
     std::string name;
-    unsigned int encounterChance; // chance ranges from 0 - 1
+    float encounterChance; // chance ranges from 0 - 1
     bool wasVisited; // flag indicating if a square has been visited
     Room<T>* prev;
     Room<T>* next;
@@ -19,7 +19,7 @@ public:
 
     void setId(const unsigned int& id);
     void setName (const std::string& name);
-    void setEncounterChance(const unsigned int& encounterChance);
+    void setEncounterChance(const float& encounterChance);
     void setWasVisited(bool& wasVisited);
     void setPrev(Room<T>* prev);
     void setNext(Room<T>* next);
@@ -27,7 +27,7 @@ public:
 
     unsigned int getId();
     std::string getName();
-    unsigned int getEncounterChance();
+    float getEncounterChance();
     bool getWasVisited();
     Room<T>* getPrev();
     Room<T>* getNext();
